[FIX] sign: fix sign alignment and consistency issue
Before this commit:
In the sign module, there are inconsistencies with the alignment of texts and
icons,  in both the portal and backend interfaces.

After this commit:
This fix addresses alignment issues and ensures consistency between the portal
and backend interfaces. Now, texts and icons in the sign appear aligned and
consistent across both interfaces.

task-3883649

closes odoo/enterprise#66732

X-original-commit: e8b583ce9abc2616d6c186cc09883f4a93777340
Signed-off-by: Kamlesh Pathekar (kpt) <kpt@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>