[FIX] sign: fix sign alignment and consistency issue
Before this commit:
In the sign module, there are inconsistencies with the alignment of texts and
icons,  in both the portal and backend interfaces.

After this commit:
This fix addresses alignment issues and ensures consistency between the portal
and backend interfaces. Now, texts and icons in the sign appear aligned and
consistent across both interfaces.

task-3883649

closes odoo/enterprise#63983

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>