[FIX] industry_fsm(_report): hide fsm section in project form
When `industry_fsm` is installed, the form of `project.project`
contains a section `Field Service` with only one setting (`is_fsm`).
This option is only visible in only visible in debug mode,
so the whole section should be displayed only under this conditon.
=> The group `base.group_no_one` has been added to the section.

When `industry_fsm_report` is installed, another setting appears
under that section (`worksheet_template_id`),
so the section should be displayed.
=> The group that is added in `industry_fsm` is removed then.

task-3916982

closes odoo/enterprise#67108

X-original-commit: 5426de1464c853df9dba7bf2dd8ee516c4e1663f
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>