[FIX] whatsapp: template sync process
Only a few templates are fetched when the user tries to sync all the templates
from the `WhatsApp Business Account` form view.

Problem:

The API request used to fetch all templates incorporates pagination. In the
current scenario, the next URL to the subsequent page is included in the
response, but we are not using it. Also, the limit parameter is not utilized in
our API request to restrict the number of templates on a single page.
Consequently, by default, 25 templates are received on the first page.
As a result, only templates received on the first page are being created or
updated(i.e. 25).

Solution:

The limit to fetch 200 templates on a single page is added to obtain 200
templates in the first request. However, if there are more than 200 templates,
the remaining ones will be fetched from the next page. This process will
continue in a loop until all templates are received.

Task-3903321

closes odoo/enterprise#67144

X-original-commit: 289c4566372b3d3a5eff0652087804948aa1cbfb
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>