[FIX] l10n_be_hr_payroll: fix paid time off wizard traceback
In the Payroll app with a Belgian company, we can do
Reporting > Paid Time Off Allocation.
This opens a wizard, and if we filled the department then the time off type,
the app would crash because of a terrible sql query.

Also, the query was doing unnecessary subqueries that have been removed.

Steps to reproduce:
- Install test_l10n_be_hr_payroll_account
- Set current company as Belgian company
- Go in Payroll > Reporting > Paid Time Off Allocation
- Set a Department in the wizard
- Select a Time Off Type

What is the bug:
- Traceback when setting the Time Off Type: https://pastebin.com/dMvqACJj

Expected behavior:
The Time Off Type should be set without traceback.

Fix:
Redo the SQL query from scratch.

closes odoo/enterprise#63723

Task: 3878904
X-original-commit: 788301cf0abe9ed097f993a4b6143fd5b2522051
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Arthur Pierrot (arpi) <arpi@odoo.com>