[IMP] website_appointment: allow customizing appointment display
Appointment Type:
In order to customize the appointment views on the website and the mail
template, we need to add some fields to the appointment type model:
- hide_duration
- hide_timezone

Appointment Templates:
With the two fields created above, we add some t-if condition to show or
not some informations on the templates views.

website_appointment:
Add a new snippet to customize the page related to an appointment
type that contains 4 checkboxes:
- Show Timezone
- Show Duration
- Show User/Resource Pictures
- Allow Guests
When triggered, these options will modify the record of the
appointment type related to the view.

As we want this snippet to be shown during the whole flow of the
appointment registration, we created a common CSS class :
o_wappointment_type_options
This class will be used to show the snippet in each view of the flow.

The snippet is associated to a JS snippet that will handle the calls
to the ORM in order to update the record of the appointment type.

Mail Templates are also updated to show hide timezone and duration depending
on the appointment type configuration.

task-3568931

closes odoo/enterprise#60819

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>