[FIX] l10n_mx_edi: use current currency rate for replacement invoice
Steps to reproduce:
- Install Accounting and l10n_mx_edi
- Switch to a Mexican company (e.g. ESCUALA KEMPER URGATE)
- Activate another currency (e.g. USD)
- In Accounting settings, run the "Automatic Currency Rates" feature ([MX] Bank of Mexico)
- Add a different currency rate for USD for yesterday
- Create an invoice:
  * Customer: [a Mexican customer] (e.g. INMOBILIARIA CVA)
  * Currency: [USD]
  * Invoice Date: [yesterday]
  * Invoice Lines:
    - Product: [any product with UNSPSC Category set]
    - Price: [any]
    - Taxes: [any]
- Confirm the invoice
- Generate CFDI via "Send & Print" button
- Request Cancel:
  * Reason: 01- Invoice issued with errors (with related document)
- Create Replacement Invoice

Issue:
A UserError is raised while checking the balance of the replacement invoice
because the total of debits and the total of credits are different.

Cause:
The lines for the replacement invoice are just copied, keeping the values of
the original invoice lines for many fields (such as "balance", "discount_balance"
and "tax_base_amount").
Product-type lines do not have "balance" set in the copied data.
In this case, "balance" is computed by using the current currency rate when the
replacement invoice is created, which is correct.
The other lines are populated with the copied "balances" that correspond to the
previous currency rate, which is not correct.
This leads to a mix of balances linked to different currency rates.

These fields are in the invoice currency and should be computed with the current
currency rate as the replacement invoice is created with the present day as invoice
date.

Solution:
Remove non product-type lines from the copied data to force them to recompute with
the current currency rate.

opw-3863010
opw-3902683

closes odoo/enterprise#62373

Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>