[IMP] project_enterprise: add highlight feature in task calendar view
Before this commit, when the user is in the form view of a task with
some conflicts with other ones and he would like to have an overview
to see the tasks in conflict in the planning by clicking on the button
displayed in the warning, he can see the tasks in conflict are
highlighted in the gantt view and the other tasks have a little opacity
to be able to see the whole planning and not just the tasks in
conflict. The problem is the highlighting is only available in the gantt
view and not in the calendar view.

This commit adds the highlighting feature in the calendar view to be
able to see the whole planning in the calendar view as in the gantt
view.

task-3514831

closes odoo/enterprise#48619

Related: odoo/odoo#157054
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>