[ADD] l10n_ke_edi_oscu(_stock/_mrp): Kenya eTims OSCU
See also https://www.odoo.com/documentation/17.0/applications/finance/fiscal_localizations/kenya.html

The previous implementation was through a fiscal device
(Tremol: l10n_ke_edi_tremol), but it will be deprecated
and with this implementation we also avoid the hassle of
needing a fiscal device.

On the other side, however, this implementation requires
far more information, even from stock and mrp.  The
stock module even depends on sale and purchase, because
we need the link between pickings and invoices as
when we send the invoices we also need to send the
 corresponding stock movement.

The mrp module is for sending BoMs.

A lot of the buttons are in debug mode, because they
are needed for automated testing e.g. and might be/come in handy,
but are not necessary for the business flows in itself.

We decided to implement OSCU as it provides a direct
connection to the web service.  (There is also a VSCU mode
which we do not support in standard, which requires you
to install libraries as a virtual, but allows for offline mode)

opw-3603762

closes odoo/enterprise#66841

X-original-commit: 3577a34f6cfb4d03918ca9a807c52b8671f1c27b
Related: odoo/odoo#173548
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Co-authored-by: Daniel Kosky <dako@odoo.com>