[FIX] sign: multi-line item losing focus when hitting enter
Before this commit, when hitting enter while typing in multi-line text,
The cursor would move to the next sign item. But it should simply add a new line and remain focused on the multi-line item.

This commit aims to fix this issue by preventing jumping to the next item when being at multi-line.

closes odoo/enterprise#63490

Task: 3910934
X-original-commit: 01ec1b52e2d947e5360531501aafe8cd009b1f78
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>