[FIX] sign: multi-line item losing focus when hitting enter
Before this commit, when hitting enter while typing in multi-line text,
The cursor would move to the next sign item. But it should simply add a new line and remain focused on the multi-line item.

This commit aims to fix this issue by preventing jumping to the next item when being at multi-line.

closes odoo/enterprise#63817

Task: 3910934
X-original-commit: ec1f50d305e07e0b32a151598bda16924b1166ec
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>