[IMP] sale_timesheet_enterprise{_holidays}: set the billing target rate in % on the employee form
This commit introduces the following changes:
- Remove billing rate target from the company settings.
- Instead, replace the billing time target on the employee form
by a billing rate target.
- Change the computation of the billable rate in the leaderboard by using
the employee working calendar (and not the billing time target anymore).
- Change the comparison made to know if an employee has reached its target
by using the billing rate target set for the employee (and not
the target rate defined in the company settings anymore).

The rationale is that as each month is different, so is the working calendar
of employees. To take the variations between months into account, we
now compute the billable rate of employees as such:
Current number of billable hours of the employee / Number of hours currently
worked by the employee according to his working calendar

task-3609085

closes odoo/enterprise#54238

Related: odoo/upgrade#5535
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>