[FIX] l10n_cl_edi_pos: Avoid search with falsy domain
Previously, to avoid loading records when loading data into
point_of_sale we used a domain with a falsy value `[('id', '=', False)]`
but this required a query to the database.

Now we return False for the domain and if the domain is false we don't
search for records. no record search.

closes odoo/enterprise#62422

Related: odoo/odoo#165344
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>