[IMP] sign: Improve sign onboarding tour
Before this commit, users needed to click on the signature field
to move to the next step.

After this commit, If a user drags the signature item and drops it
into the iframe document, it will move to the next step without
the need for clicking.

task-3925591

closes odoo/enterprise#62447

Related: odoo/odoo#165395
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>