[FIX] planning: change default shift visibility
Issue
Currently, internals users are shown all the open shifts when they arrive on
the planning app, even if the shift's role is not one of the user's role.

Fix
- If the employee doesn't have any roles, show all open shifts.
- If the employee has role(s), show the open shifts with that role or no
role.
To do this, we modify the open shifts filter (which is set by default for
internal users) to only show open shifts matching the user's role in the
default Gantt view.

opw-3888507

closes odoo/enterprise#65472

X-original-commit: 773457b382eeda10cf1001007821b2e80eabe17a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>