[FIX] pos_preparation_display: refresh orders if all PoS allowed
A domain is used to filter which preparation displays should be refreshed
when an order arrives.
The current domain only filters the preparation displays that are linked to
one or more PoS. If no specific PoS is linked to the preparation display,
the domain will filter out the preparation display and the user will have to
manually refresh the preparation display to see the new order.

opw-3756416

closes odoo/enterprise#62430

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>