[FIX] mrp_workorder: round consumed qty in Shop Floor view
Issue:

A quantity of `3` gets displayed as `3.00...004`.

Steps to reproduce:

- Settings > Technical > Database Structure > Decimal Accuracy
- Change the `Digits` of `Product Unit of measure` to 5
- Create a storable product P1 with a BOM:
    - In the operation tab: one operation in `Assembly line 1`
    - In the component tab: 1 unit of P2 consumed in the operation
- Create and confirm an MO for 3 units of P1
- Go to the shopfloor module, remove the MO ready filter
Two issues can be noticed:
- Click on Register Production
A pop up opens. Before validating it:
The displayed producing quantity of P1 and the quantity of P2 that
should be consumed by the operation is`3.00...004`. After validation,
the produced quantity is rounded. However, if you click on the cross
the rounding error will remain. In addition, the second rounding issue
will remain even after validation of the consumed qty.

Cause of the issue:

Our float utils introduce tiny rounding errors, especially at higher
precisions. In our case, these errors are introduced during the
call of the write method of the `web_save`.

Fix:

We eliminate the rounding errors for the display in the same way as
initially ptoposed by:
https://github.com/odoo/enterprise/pull/58880
see commit 1c2f006:
https://github.com/odoo/enterprise/commit/1c2f0068d319917bb4af163396d0861a1ff6c9e1
and adapt it to also apply to the raw_move_ids of the MO.

opw-3861892

closes odoo/enterprise#62510

Signed-off-by: Steve Van Essche <svs@odoo.com>