[FIX] sale_planning: ensure form view used in plan dialog is gantt form
Before this commit, when the user clicks on a gantt cell in the planning
gantt view, the `Plan` dialog is displayed and if he clicks on `New`
button, the form view dialog is not the same form view displayed when we
click on the new view in the gantt view.

This commit fixes the issue by given the right form view to the form
view dialog when the user clicks on the `New` button inside the plan
dialog.

task-3636218

closes odoo/enterprise#55625

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>