[FIX] documents_hr: fix `documents_hr_{documents,folder}` deletion
The record `documents_hr_contract.documents_hr_documents_contracts` has
forcecreate=1 and there is a chain of dependency of these 3 records:

1. `documents_hr_contract.documents_hr_documents_contracts` --> 2.
2. `documents_hr.documents_hr_documents`                    --> 3.
3. `documents_hr.documents_hr_folder`

But for the latter 2 records forcecreate is set `0`. And all of these
records are set `noupdate=1`. That is why during upgrade the latter 2
records will be skipped from creation if they are missing in db, but
it will try to create the 1st one because of that forcecrete=1.
This will cause dependency issue if the db is missing the other
2 records.

For solving the issue we set `forcecreate=1` to the latter 2 records.

Steps to reproduce:
1. Install documents_hr, documents_hr_contract in version `16.0`
2. Delete the document facet `documents_hr.documents_hr_documents`
3. Delete the document folder `documents_hr.documents_hr_folder`
4. Try to upgrade to `17.0`

You will see similar traceback to this:

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/17.0/odoo/modules/registry.py", line 113, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)

  ...
  ...

  File "/home/odoo/src/odoo/17.0/odoo/tools/convert.py", line 693, in convert_xml_import
    obj.parse(doc.getroot())
  File "/home/odoo/src/odoo/17.0/odoo/tools/convert.py", line 613, in parse
    self._tag_root(de)
  File "/home/odoo/src/odoo/17.0/odoo/tools/convert.py", line 556, in _tag_root
    f(rec)
  File "/home/odoo/src/odoo/17.0/odoo/tools/convert.py", line 569, in _tag_root
    raise ParseError('while parsing %s:%s, somewhere inside\n%s' % (
odoo.tools.convert.ParseError: while parsing /home/odoo/src/enterprise/17.0/documents_hr_contract/data/documents_tag_data.xml:10, somewhere inside
<record id="document_tag_signature_request" model="documents.tag" forcecreate="1">
            <field name="name">Signature Request</field>
            <field name="facet_id" ref="documents_hr.documents_hr_documents"/>
            <field name="sequence">15</field>
        </record>
```

closes odoo/enterprise#62584

X-original-commit: fe86c163b98f483e5e8f6258b03e320b1e5c9f75
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>