[FIX] knowledge: fix inconsistent id in article url for portal users
When a portal user opens Knowledge, the user loads a page that loads
the backend assets. The client script will then read the article id from
the session and execute an action to open that article. When the action
is executed, the action service will update the url params and add a
`#id=<article_id>` entry to the url params.

When the user opens another article, the system will execute a new action
and the action service will update the url param `#id`. However, it will
not update the id specified in the route nor the id stored in the session.

This makes the url inconsistent and cause the following issue: When the
user reloads the page, the server will copy the article id specified in
the route (`/<article_id>`) into the session and ignore the id specified
in the url params (`#id=<article_id>`). The url parameter will not be
transmitted to the server. The client will then load the wrong article.

Steps to reproduce the issue:
1. Open an article
=> The url becomes `/knowledge/article/1#id=1`
2. Click on an article from the sidebar
=> The url becomes `/knowledge/article/1#id=2`
3. Reload the page (i.e: Press F5)
=> The system reloads the article with the id 1

TO BE: When the user reloads the page, the system should reload the last
article opened (article 2) and not the one initially loaded (article 1).

To fix the issue, we will now: 1) Redirect the portal user to a new route
that simply moves the article id from the url to the url params and 2)
make sure that the client reads the article id from the url params
(#id=<article_id>) and not from the session. This should ensure that the
url, url params and session remain consistent.

When a portal user accesses the `/knowledge/article/1` route, we will
now redirect the user to `/knowledge/article#id=1`. The client script
will then load the article specified in the url params. If the user
opens another article from the sidebar, the action service will update
the `id` url parameter.

With those changes, the article id is no longer duplicated in the url.
If the client reloads the page, the client script should load the correct
article as the script loads the article using the id url parameter and
that parameter is updated when the user opens a new article.

task-3743416

Part-of: odoo/enterprise#62861