[FIX] stock_barcode: scan source location
The commit [1] fixes the fact than users could scan any locations as the
source, even if the scanned location is not related to the picking's
source location.
But for immediate transfer, it is not very wise to limited which
location the user can scan because they have no way to set source or
destination for picking created on the fly from the Barcode app.
This commit fixes this issue.

How to reproduce:
- Inventory -> Settings -> Enable "Multi-Step Routes";
- Go to your warehouse and choose 3 steps for incoming and/or outgoing
  shipments;
- Go to the Barcode App > Operations > Internal Transfers > New;
- In the newly created transfer, try to scan who is not a sublocation of
  WH/Stock (WH/input for example, barcode: WH-INPUT)
  -> An error message is displayed because you are not allowed to scan
     this location.

The exact same issue happens for destination too since [2].

Also, fix a minor unrelated bug where locations on the barcode line are
not correctly written when there name include the picking's defaut
location's name (eg.: WH/Stock 2 when the default location is WH/Stock.)

OPW-3843358

[1] 0c84f2d86b982d1b244622f2d86ca474f0ba496b
[2] 61d378b0a49ba734cff84a39aa2693fe4226225d

closes odoo/enterprise#62733

X-original-commit: a6d14aff47cfbf4253a86b975d032b63e2281611
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>