[FIX] planning: make form date reflect selected block in calendar
Steps to reproduce the issue:

1. Go to _Planning > My Planning_ and switch to the Weekly Calendar view
2. Plan a new Event from 5.30 AM to 7 PM in the Calendar
3. When the creation Form opens, the Date does not reflect the slot you were creating

- Also happens in Gantt view:
    1. Make sure the company's schedule starts or ends between two hours (e.g.: 8.15 am)
    2. Click on the '+' button of the corresponding cell in the Daily Gantt view
    3. The Date will be cut to the start or end of the schedule.

Explanation:

Before opening `planning.planning_view_form`, the client calls for `onchange` which will then call for `default_get` that will align `start_datetime` and `end_datetime` with the work hours of `res.company.resource_calendar_id` in every situation where no resource is specified.
https://github.com/odoo/enterprise/blob/ff4e39a09399f46790c0ad45bc9f9af4e4cd44d1/planning/models/planning.py#L625-L635

Suggested fix:

`start_datetime` and `end_datetime` should not change only in Weekly and Daily Calendar views and on Daily Gantt View.
The key should not remain in the Context of the Action to avoid unintended behaviours, adding it in the Context of the Event instead.

opw-3829635

closes odoo/enterprise#62517

X-original-commit: 581eca78f7a3e0cfbb93c28dd53b4f9d9ff9d614
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>