[FIX] knowledge: restore kanban group background
This commit fixes an issue that was brought by the redesign of Odoo, Milk.
This redesign modified a scss variable for kanban views inside of form views
which removed the background of the kanban view's headers.

Now a new css rule has been added to the kanban embedded view so that the
background comes back to its original value inside of Knowledge.

task-3930157

closes odoo/enterprise#62758

X-original-commit: 51e3e2f5c0e0c158c263d7e92af0937e278c64f0
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>