[FIX] planning: fix test tour
Before this commit, the #55625 altered the planning_tests_tours to
change Aaron into Mitchell Admin for the resource assigned to the shift
created inside that tour. The problem is `Mitchell Admin` is a demo data
and so, the tour will create a new resource when we have no demo data
inside our DB. If the changes made by this PR in that tour are reverted
then another error is raised due to a timezone mismatch.

This commit reverts the changes made by #55625 inside
planning_tests_tours and change the timezone of `Aaron` resource created
inside the python test.

closes odoo/enterprise#62567

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>