[IMP] appointment: add route and update xml
This commit adds the route for checking the synchronization status
of calendar integration modules in order to fix the current QUnit
tests. Additionaly, a small css change in the override of the
'Share Availabities' button is made in order to fit the new
synchronizations buttons on the header alongside it.

related: odoo/odoo#150186
task-3570379

closes odoo/enterprise#60155

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>