[FIX] account_reports: race condition in sortable test case
closes odoo/enterprise#62611

Issue: when the account_reports tour test is run it sometimes fails with some report line not containing the correct value. The root cause of this issue is that the triggers for 2 of the test steps check the content of lines that are excluded from sorting. This could result in the test step triggering before sorting of the report is completed, which will in turn fail the testcase.
Fix: while testing the sorting mechanism only trigger on lines that are included in the sort. This ensures the next test step is triggered after sorting is completed.
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>