[FIX] account_reports: break word if overflow in PDF when RTL
Steps to reproduce:
- Install account_reports
- Install a RTL language (e.g. Arabic)
- Set the RTL language on the current user
- Set the currency of the company to SR
- Go to "Accounting / Reporting / Partner Ledger"
- Make sure there are some partners with long name (but less than 50 chars)
- Print PDF

Issue:
In RTL, the partner column overflows outside of the page width.

Cause:
The break-word is only activated when the name has more than 50 characters.
In RTL, the cell to display the partner is smaller.

Solution:
Do not use a minimum length of 50 characters to activate "break-word" in RTL.

opw-3888122

closes odoo/enterprise#63045

X-original-commit: ac83b1a4a455708f33f6a29918c4e7212b59c464
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>