[FIX] hr_contract_salary: prevent traceback when clicking on contracts
When user clicks on contracts smart button in application and if application has
two contracts (one archived and one active), a traceback will appear.

Steps to reproduce the error:
- Install "hr_contract_salary" module
- Go to Recruitment > Applications > All Applications > Create new application >
  Generate Offer > Select Contract Template > Send By Email > Send
- Click on Offers > Salary Configurator > Fill all required fields >
  Review Contract & Sign > Validate & Send Completed Document
- Open that application > Contracts > Create new contract > Save
- Archived one contract > Click on Contract smart button

Error: A traceback appears:
"Expected singleton: hr.contract(27, 29)"

https://github.com/odoo/enterprise/blob/504b3c66a9a8d142b89f9c873675ec65fdb43dd0/hr_contract_salary/models/hr_applicant.py#L59
Here when application has two contracts (one is archived and one is active),
So when it tries to access the id of 2 records.
It will lead to the above traceback.

sentry-5350130147

closes odoo/enterprise#63294

X-original-commit: e9a21a3b92f15bd45d403c4dbba1a8d47ccd1fab
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>