[IMP] knowledge : update a query test
This commit modifies the number of expected queries in a performance
test of knowledge. The increase in queries results from the
implementation of the embedded actions task, which requires one additional
query to retrieve the correct values. Consequently, the expected number
of queries rises from 19 to 20.

taskid:3750721

closes odoo/enterprise#61388

Related: odoo/odoo#163184
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>