[FIX] l10n_ec_edi: prevent an error while installing a 'l10n_ec_edi' module
Currently, an error occurs when installing a 'l10n_ec_edi' module.

Step to produce:

- Create a company and set its country 'Ecuador'.
- And install the 'l10n_ec_edi' module.

ValueError: External ID not found in the system:
account.1_ec_sale_withhold_tax_base

An error occurs when the system tries to retrieve a 'Sales Tax Base Account'
through account view id at [1], but the view id is not available.

To resolve this issue, add 'raise_if_not_found=False' so that if the external id
of view is not found, the 'Sales Tax Base Account' gets an empty record instead.

sentry-5183479269

closes odoo/enterprise#63910

X-original-commit: 63ad9422e48b6cc93ab9f0e92d4e79671edf4856
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>