[FIX] account_accountant,*: make bank rec tab titles translatable
*account_accountant_batch_payment, sale_account_accountant

Currently the tab titles of the bank reconciliation widget are not
translatable. This is due to the fact that attributes of OWL components
(including `t-set-slot` elements) are not considered translatable.

The solution is to extract these translatable terms in a variable
outside of the `t-set-slot`'s parent component so it would become
translatable and available inside the right slots.

The issue was already partially fixed in [1] in 17.0.

[1]: https://github.com/odoo/enterprise/commit/161163928bc4a18700be1b69b5276cb3b8490e45

opw-3881861

closes odoo/enterprise#62668

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>