[FIX] knowledge: fix crash on retry comments notifs
This commit fixes an issue with comments notifications in Knowledge.
When a user tags another user in a Knowledge comment, a mail should be sent.
When an error occurs the user have the possibility to retry to send the
notification.
Retrying this action triggers an error with threads because the msg_vals key
doesn't exist when trying to access it inside `_notify_thread_by_email`.

The fix is to use the get function of Python dict so that if the msg_vals aren't
set, no error is triggered.

task-3933062

closes odoo/enterprise#62649

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>