[MOV] sign: bidi get_displayed_text
The function get_displayed_text is used for text processing (reshaping
and redirecting) mainly for printing on pdf.
The functino was first introduced in the sign module, original PR: odoo/enterprise#36315

This commit aims to move this function to odoo.tools.pdf and rename it to 'reshape_text'.

Task: 3494707
Part-of: odoo/enterprise#62583
Related: odoo/odoo#165774
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>