[FIX] account_reports: fix general ledger not receiving options
This issue arises after having visited the "General Ledger" at least
once.

When clicking on the "General Ledger" part of the "The current balance in
the General Ledger didn't match the balance of your last bank statement,
leading to an unexplained difference of [...]" message in "Bank
Reconciliation", options such as the currency unit weren't properly
transferred.

Steps to reproduce original issue:

- Install `account_reports`
- Go to: Accounting > Reporting > Audit Reports > General Ledger
- Go back
- Create, confirm & reconcile an invoice
- Reset invoice to draft
- In "Accounting", go to "Dashboard"
- Click on the 3 vertical dots to the right of the "Bank" card
- Click on "Reconciliation Report"
- In "Options: Posted Entries Only", select "Include Draft Entries"
- Click on "General Ledger" in the first yellow warning
- Draft entries won't be included

task-3928454

closes odoo/enterprise#63032

X-original-commit: 3bff8b10c3f7d318810c2148439d49c2ec65c6f3
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Solal Vignal (sovi) <sovi@odoo.com>