[FIX] documents: inspector behavior in mobile
Steps to reproduce:
1. Go to documents on mobile
2. Click on doc thumbnail
3. The inspector open where preview should open

When the preview is open, click on an arrow to see the previous or the next doc
1. The inspector opens again where the previous or next should

Technical Reason:
In 'documents_inspector.js' file, a condition set only open the inspector if
there is only one selected element, so every time it get executed and opens
the inspector.

After this commit:
1. Click on a doc thumbnail, the preview will open.
2. When the preview is open, click on an arrow to see the previous or the next
doc will not open the inspector.

Task-3919285

closes odoo/enterprise#72677

X-original-commit: f3c682224180bcbec2dbefc96b7ea1251b506659
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Astik Singh (assi) <assi@odoo.com>