[PERF] spreadsheet_edition: don't export data when leaving
Currently, when leaving a spreadsheet document, the data is exported and
given to a callback. Exporting the data can take several seconds on large
spreadsheets though. It slows the user experience.

Currently, this exported data is only used for Documents to check if
the spreadsheet has multiple sheets. We don't need to export the full data,
we can just read the spreadsheet model.

closes odoo/enterprise#62701

Task: 3934885
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>