[FIX] hr_contract_salary: recompute yearly cost with holidays
When the yearly cost is recomputed, beecause the change of a benefit
re-triggered it, it is badly recomputed if the contract contains extra
time off.

To avoid this issue, we us the function _get_yearly_cost with inverse
param to true that correctly handle the value of the extra time off in
the yearly cost.

We've also modified the way '_get_yearly_cost' is computed as it when we
compute wage_with_holidays we need to get the final_yearly_costs, and
the final_yearly_costs calls '_get_yearly_cost' which was needing the
wage_with_holidays wich is the one we try to compute.

closes odoo/enterprise#62704

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>