[FIX] account_accountant: Fix badly linked exchange_move_id on bank rec widget
Steps to reproduce:
- Create an invoice with a foreign currency of rate 1:3:
120 EUR = 40 USD
- Create a statement line with a foreign currency of rate 1:2:
120 EUR = 60 USD
- Reconcile them both using the bank reco widget:
An exchange diff of 20 USD will be created.

At this point, the invoice and the statement line are reconciled all together.
The exchange difference is reconciled with the statement line (lowest rate).

The point is both account.partial.reconcile are linked to the exchange diff.
That's what will cause the problem describe below.

When the invoice is reset to draft:
'button_draft' removes the reconciliation so the partial linking the invoice with the statement line is deleted.
Since an exchange diff is attached to the account.partial.reconcile through the exchange_move_id field, we
create a reversal for the exchange diff to neutralize its accounting items.
During the creation of the reversal, we remove the reconciliation so the partial linking the exchange difference
with the statement line is deleted.
Since the exchange diff is wrongly attached to this partial too, again, the code tries to reverse the exchange
difference.
At the end, a double reversal of the exchange diff has been created.

closes odoo/enterprise#62968

Opw: 3869770
X-original-commit: 41857e41c5940b891c2a468f6e2358843088bcc8
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>