[FIX] l10n_lu_reports: correct context of pl_abr action
The link from 'Balance Sheet (abridged)' to 'Profit & Loss (abridged)' raises an error.

Steps to reproduce :
- Install the l10n_lu_reports module
- Change the company to a Luxembourgish comppany
- Go to Accounting > Reporting > Statement Reports > Balance Sheet
- Change the report to 'Balance Sheet (abridged) (LU)'
- In the report click on the link 'VI. Profit or loss for the financial year'

Cause:
The action triggered by this link has not he right formatted context. It should have the value 'report_id' but has 'model' and 'id' instead. This raises an error in execute_action when trying to read 'report_id' on this action.

Solution:
Remove 'model' in the context and rename 'id' to 'report_id'.

opw-3912348

closes odoo/enterprise#62799

X-original-commit: a6cc2a8c58a7e2f62493626fa9ba3ae857a81694
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>