[IMP] stock_barcode: sign delivery orders
Added in this commit
1. If the user has `stock.group_stock_sign_delivery` activated,
a `Sign` button is added to the gear context menu for `outgoing`
pickings.
2. When the `Sign` button is pressed, a dialog box is opened to
take the signature drawing input from the user. `partner_id` is supplied
as a default name for convenience too. The dialog box approximately
replicates the functionality of `SignatureWidget` in the backend.
3. Signatures are checked before validation for `outgoing` pickings:
if they are not supplied, the sign dialog is opened. After the signature
is supplied from the user, picking gets validated automatically.
4. Batch pickings are not included in this commit.

task-3601184

closes odoo/enterprise#64211

Signed-off-by: Steve Van Essche <svs@odoo.com>