[FIX] account_reports: Journal Report: avoid the report becoming too large when unfolding lines
When unfolding the journals, a lot of empty, unused space was left around the values. This lead to the report easily becoming too large for the screen, causing the display of an horizontal scrollbar, and hence degrading the user experience. We fix that by removing any notion of minimal width for this report specifically.

closes odoo/enterprise#62718

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>