[FIX] account_payment: allow duplicate acquirer per company
The payment acquirer must always be linked to a single journal, one per company. It's currently an hard constraint
So after the duplication, both are always linked to the same journal but only one appear on it.
This "fix" is improving that allowing the duplication of an existing acquirer and to be able to link them to diffe

opw-3704407

closes odoo/enterprise#63214

X-original-commit: d5368053f377d4711e6d722a1c99146588b416ba
Related: odoo/odoo#166745
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>