[FIX] account_reports: properly handle non-float figure types on prefix group lines
To reproduce the issue, on a runbot with the demo data:

1) Setup a prefix group threshold of 2 on the Partner Ledger
2) Open the Partner Ledger

==> The lines created for the prefix groups show "0.0" in the columns supposed to contain non-number values. Instead, these columns should contain empty values on those lines.

The bug originates in the common helper called to generate the prefix group lines, so it does not only impact the Partner Ledger. We fix it and modify the test a little bit to check that behavior as well.

closes odoo/enterprise#63052

X-original-commit: 9a5749d0e635dd023b5365b82da771b1f0fbae8e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>