[IMP] knowledge: add article finder
Knowledge is a wonderful tool for collecting, organising and sharing
documents with colleagues. However, Knowledge lacks an effective search
feature allowing people to search within the content of their documents.

The existing search feature only searches for documents with a specific
title. If the users don't know which document contains the information
they need, they can waste time reading a bunch of articles before finding
what they want. This can be time consuming and frustrating when there
are a lot of documents in the knowledge base and when the articles are
not well organised.

To help people to quickly find the information they need in the knowledge
base, we will improves the command palette search algorithm. The search
algorithm will now be able to find documents containing the search terms
inputted by the user.

Technical
---

The new search algorithm uses the PostgreSQL's built-in full-text search
feature. This feature will do all the heavy lifting for us: It will be able
to strip the html tags from a document, quickly find article matching with
the given search terms, highlight the matching terms in the document, etc.

To index the documents, we will define a custom text configuration that
is independent of the user's language: It will not discard any words and
will not apply any stemming methods to the documents and the search query.
It will also ignore the HTML tags stored in the article body.

To reduce the query runtime, the search algorithm limits the number of
candidates to consider (see: the "cut_off" parameter) and pre-selects
relevant candidates incrementally:

The algorithm first selects articles matching with the title and the body
of the article. If there are not enough matches, the algorithm will select
articles matching with the title only. If there are still not enough matches,
the algorithm will select the articles matching with the body only.

After that, the algorithm ranks the selected articles using a scoring
function and returns the most relevant ones. This should ensure that the
results we get are relevant and that the query response time is acceptable.

Benchmark
---

To evaluate the efficiency of the new search algorithm, we loaded in
Knowledge 100, 1000 and 10 000 articles coming from an online
encyclopaedia using a populate script. The script will nest these
articles under one another to form a tree with a branching factor of 5.
The articles will be written in English and will be placed in the
workspace of Knowledge.

We then randomly selected 100 different words from the loaded articles
and we searched them using the new search algorithm. We then measure the
query response time and we calculate the average, the standard deviation,
the minimum and the maximum response time.

Response time:

- Admin user:

|     N |       AVG | Std Dev |  Minimum |   Maximum |
|-------|-----------|---------|----------|-----------|
|   100 |  188.92ms |   172.5 |   18.4ms |  540.31ms |
|  1000 |  406.56ms |  316.31 |  28.54ms | 1236.47ms |
| 10000 | 1467.31ms |  901.12 | 138.56ms | 3670.40ms |

- Demo user:

|     N |       AVG | Std Dev |  Minimum |   Maximum |
|-------|-----------|---------|----------|-----------|
|   100 |     184ms |  165.41 |  25.86ms |  571.79ms |
|  1000 |  457.69ms |  275.76 |  76.02ms | 1456.63ms |
| 10000 | 1504.90ms |  662.64 | 600.57ms | 3199.80ms |

**Observation**: The response time can vary greatly depending on the
number of times the search terms appear in the documents. The more
often the search term is used, the more results there will be and the
slower the query will be.

- The minimal response time observed is when the search term appears in
the body of a single article.
- The maximal response time observed is when the search term appears in
the body of many article but does not appear frequently in the titles.

Index disk usage:

|     N |                    Index name |   Size |
|-------|-------------------------------|--------|
|   100 | knowledge_article__name_index |  229kB |
|   100 | knowledge_article__body_index | 2318kB |
|  1000 | knowledge_article__name_index | 1458kB |
|  1000 | knowledge_article__body_index | 8536kB |
| 10000 | knowledge_article__name_index | 5447kB |
| 10000 | knowledge_article__body_index |   32MB |

Column disk usage:

|     N | Column name |   Size |
|-------|-------------|--------|
| 100   |        body |  855kB |
| 1000  |        body | 6799kB |
| 10000 |        body |   71MB |

Legend:

- N: Number of articles stored in the database

Reference:

PostgreSQL documentation: https://www.postgresql.org/docs/15/textsearch.html

task-3071299

Part-of: odoo/enterprise#52106